requirements.txt
dspy-ai
pydantic
